pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "molplay"
version = "0.1.0"
description = "Protein-ligand binding affinity playground with generation demos"
readme = "README.md"
authors = [{name = "Protein Molecule AI Playground"}]
requires-python = ">=3.11"
dependencies = [
  "typer>=0.9",
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "torch>=2.1",
  "torch-geometric>=2.4",
  "rdkit-pypi>=2022.9",
  "pyarrow>=13.0",
  "tqdm>=4.66",
  "matplotlib>=3.7",
  "seaborn>=0.13"
]

[project.optional-dependencies]
tests = ["pytest>=7.4", "pytest-cov>=4.1"]

[project.scripts]
molplay-train = "molplay.train:app"
molplay-generate = "molplay.generate:app"

[tool.setuptools.package-data]
molplay = ["py.typed"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.setuptools.packages.find]
include = ["molplay*"]
